requirements.txt
# Core framework
fastapi>=0.115.13
uvicorn[standard]>=0.24.0
websockets>=12.0

# Browser automation
browser-use>=0.1.18
playwright>=1.48.0

# AI/LLM integration
openai>=1.68.0
anthropic>=0.39.0

# Environment and utilities
python-dotenv>=1.0.0
pydantic>=2.5.0
Pillow>=10.1.0

# VNC and streaming
pyvnc>=0.1.0
asyncio-mqtt>=0.16.0

# Optional dependencies
httpx>=0.26.0
aiofiles>=23.2.1